+++ layers/Engine/packages/hooks/tests/Services/HooksAPI.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PoP\Hooks\ContractImplementations;
+namespace PoP\Hooks\Services;
 
 use PoP\Hooks\HooksAPIInterface;
+
 class HooksAPI implements HooksAPIInterface
 {
     public function addFilter(string $tag, callable $function_to_add, int $priority = 10, int $accepted_args = 1): void
